requirements.txt
# Voice Talker - Зависимости проекта
# Обновлено: добавлены зависимости для потокового ASR и OpenAI 2.x


# ==================== LLM и AI ====================
# OpenAI SDK 2.x для работы с локальными моделями (LM Studio, Ollama)
openai>=1.0.0

# Загрузка переменных окружения
python-dotenv>=1.0.0

# ==================== Веб-фреймворки ====================
# FastAPI для API сервера
fastapi>=0.100.0
python-multipart>=0.0.6

# ASGI сервер
uvicorn[standard]>=0.20.0

# Flask (для совместимости с существующими модулями)
flask>=2.3.0

# WebSocket поддержка для FastAPI
websockets>=10.0

# ==================== Обработка аудио ====================
# Numpy для работы с аудио данными
numpy>=1.24.0

# Опциональные зависимости для обработки голоса
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
pydub>=0.25.1

# ==================== T-one ASR (потоковое распознавание) ====================
# Примечание: T-one требует дополнительной установки
# Можно установить из репозитория: pip install git+https://github.com/voicekit-team/T-one.git
# Или скопировать папку tone в проект

# ONNX Runtime для работы модели
onnxruntime>=1.15.0

# Protobuf - фиксированная версия для совместимости с ONNX
protobuf>=3.20.0,<5.0.0

# Transformers для загрузки моделей из HuggingFace
transformers>=4.30.0

# Huggingface Hub для загрузки моделей
huggingface-hub>=0.16.0

# CTC декодер для T-one ASR
pyctcdecode>=0.5.0

# KenLM языковая модель для beam search декодирования
# Примечание: требует cmake для сборки
kenlm @ https://github.com/kpu/kenlm/archive/master.zip

# ==================== Telegram бот ====================
python-telegram-bot>=20.0

# ==================== WhatsApp (Twilio) ====================
twilio>=8.0.0

# ==================== HTTP запросы ====================
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.24.0

# ==================== База данных ====================
sqlalchemy>=2.0.0

# ==================== Утилиты ====================
pydantic>=2.0.0
typing-extensions>=4.5.0

# ==================== Разработка и тестирование ====================
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# ==================== Опциональные зависимости ====================
# Декодирование аудио различных форматов
# miniaudio>=1.59  # Легковесный декодер аудио

# Языковая модель для улучшения распознавания
# pyctcdecode>=0.5.0  # CTC декодер с KenLM

# Офлайн распознавание речи (альтернатива T-one)
# vosk>=0.3.45

# Синтез речи
# coqui-tts>=0.20.0

# Запись аудио
# sounddevice>=0.4.0
